unified conftest for all modules
