+++ docs/source/reference/test_navdata/modules.rst
@@ -4,6 +4,5 @@ navdata
 .. toctree::
    :maxdepth: 4
 
-   conftest
    test_navdata
    test_operations
