+++ docs/source/reference/test_utils/modules.rst
@@ -4,7 +4,6 @@ utils
 .. toctree::
    :maxdepth: 4
 
-   conftest
    test_coordinates
    test_dop
    test_ephemeris_downloader
