--- docs/source/reference/test_visualizations/conftest.rst
@@ -1,7 +1,0 @@
-conftest module
-===============
-
-.. automodule:: conftest
-   :members:
-   :undoc-members:
-   :show-inheritance:
