+++ docs/source/reference/test_visualizations/modules.rst
@@ -4,7 +4,6 @@ visualizations
 .. toctree::
    :maxdepth: 4
 
-   conftest
    test_plot_map
    test_plot_metric
    test_plot_skyplot
