+++ tests/algorithms/test_fde.py
@@ -11,89 +11,19 @@ import pytest
 import numpy as np
 
 from gnss_lib_py.navdata.navdata import NavData
-from gnss_lib_py.parsers.google_decimeter import AndroidDerived2022
 from gnss_lib_py.algorithms.fde import solve_fde, evaluate_fde
-
-@pytest.fixture(name="root_path_2022")
-def fixture_root_path_2022():
-    """Location of measurements for unit test.
-
-    Returns
-    -------
-    root_path_2022 : string
-        Folder location containing measurements.
-
-    """
-    root_path = os.path.dirname(
-                os.path.dirname(
-                os.path.dirname(
-                os.path.realpath(__file__))))
-    root_path_2022 = os.path.join(root_path, 'data','unit_test','google_decimeter_2022')
-    return root_path_2022
-
-@pytest.fixture(name="derived_2022_path")
-def fixture_derived_2022_path(root_path_2022):
-    """Filepath of Android Derived measurements.
-
-    Parameters
-    ----------
-    root_path_2022 : string
-        Folder location containing measurements.
-
-    Returns
-    -------
-    derived_2022_path : string
-        Location for the unit_test Android derived 2022 measurements.
-
-    Notes
-    -----
-    Test data is a subset of the Android Raw Measurement Dataset [1]_,
-    from the 2022 Decimeter Challenge. Particularly, the
-    train/2021-04-29-MTV-2/SamsungGalaxyS20Ultra trace. The dataset
-    was retrieved from
-    https://www.kaggle.com/competitions/smartphone-decimeter-2022/data
-
-    References
-    ----------
-    .. [1] Fu, Guoyu Michael, Mohammed Khider, and Frank van Diggelen.
-        "Android Raw GNSS Measurement Datasets for Precise Positioning."
-        Proceedings of the 33rd International Technical Meeting of the
-        Satellite Division of The Institute of Navigation (ION GNSS+
-        2020). 2020.
-
-    """
-    derived_2022_path = os.path.join(root_path_2022, 'device_gnss.csv')
-    return derived_2022_path
-
-@pytest.fixture(name="derived")
-def fixture_load_derived(derived_2022_path):
-    """Load instance of AndroidDerived2022.
-
-    Parameters
-    ----------
-    derived_2022_path : pytest.fixture
-        String with location of Android derived measurement file.
-
-    Returns
-    -------
-    derived : AndroidDerived2022
-        Instance of AndroidDerived2022 for testing.
-
-    """
-    derived = AndroidDerived2022(derived_2022_path)
-    return derived
 
 @pytest.mark.parametrize('method',
                         [
                          "residual",
                          "edm",
                         ])
-def test_solve_fde(derived, method):
+def test_solve_fde(derived_2022, method):
     """Test residual-based FDE.
 
     Parameters
     ----------
-    derived : AndroidDerived2022
+    derived_2022 : AndroidDerived2022
         Instance of AndroidDerived2022 for testing.
     method : string
         Method for fault detection and exclusion.
@@ -101,23 +31,23 @@ def test_solve_fde(derived, method):
     """
 
     # test without removing outliers
-    navdata = derived.copy()
+    navdata = derived_2022.copy()
     navdata = solve_fde(navdata, method=method)
     assert "fault_" + method in navdata.rows
 
     # max thresholds shouldn't remove any
-    navdata = derived.copy()
+    navdata = derived_2022.copy()
     navdata = solve_fde(navdata, threshold=np.inf, method=method)
     assert sum(navdata.where("fault_" + method,1)["fault_" + method]) == 0
 
     # min threshold should remove most all
-    navdata = derived.copy()
+    navdata = derived_2022.copy()
     navdata = solve_fde(navdata, threshold=-np.inf, method=method)
     print(sum(navdata.where("fault_" + method,1)["fault_" + method]))
     assert len(navdata.where("fault_" + method,0)) == 24
     num_unknown = len(navdata.where("fault_" + method,2))
 
-    navdata = derived.copy()
+    navdata = derived_2022.copy()
     original_length = len(navdata)
     navdata = solve_fde(navdata,
                         threshold=-np.inf,
@@ -130,18 +60,18 @@ def test_solve_fde(derived, method):
                                   np.array([0]))
     assert len(navdata) == original_length - num_unknown - 6
 
-def test_fde_fails(derived):
+def test_fde_fails(derived_2022):
     """Test that solve_fde fails when it should.
 
     Parameters
     ----------
-    derived : AndroidDerived2022
+    derived_2022 : AndroidDerived2022
         Instance of AndroidDerived2022 for testing.
 
     """
 
     with pytest.raises(ValueError) as excinfo:
-        solve_fde(derived, method="perfect_method")
+        solve_fde(derived_2022, method="perfect_method")
     assert "invalid method" in str(excinfo.value)
 
 @pytest.mark.parametrize('method',
@@ -149,19 +79,19 @@ def test_fde_fails(derived):
                          "residual",
                          "edm",
                         ])
-def test_evaluate_fde(derived, method):
+def test_evaluate_fde(derived_2022, method):
     """Evaluate FDE methods.
 
     Parameters
     ----------
-    derived : AndroidDerived2022
+    derived_2022 : AndroidDerived2022
         Instance of AndroidDerived2022 for testing.
     method : string
         Method for fault detection and exclusion.
 
     """
 
-    navdata = derived.copy()
+    navdata = derived_2022.copy()
     evaluate_fde(navdata,
                  method=method,
                  fault_truth_row="MultipathIndicator",
