+++ tests/algorithms/test_residuals.py
@@ -15,113 +15,51 @@ from gnss_lib_py.parsers.google_decimeter import AndroidDerived2021
 from gnss_lib_py.navdata.navdata import NavData
 from gnss_lib_py.algorithms.residuals import solve_residuals
 
-@pytest.fixture(name="root_path")
-def fixture_root_path():
-    """Location of measurements for unit test
-
-    Returns
-    -------
-    root_path : string
-        Folder location containing measurements
-    """
-    root_path = os.path.dirname(
-                os.path.dirname(
-                os.path.dirname(
-                os.path.realpath(__file__))))
-    root_path = os.path.join(root_path, 'data/unit_test/google_decimeter_2021')
-    return root_path
-
-
-@pytest.fixture(name="derived_path")
-def fixture_derived_path(root_path):
-    """Filepath of Android Derived measurements
-
-    Returns
-    -------
-    derived_path : string
-        Location for the unit_test Android derived measurements
-
-    Notes
-    -----
-    Test data is a subset of the Android Raw Measurement Dataset [1]_,
-    particularly the train/2020-05-14-US-MTV-1/Pixel4 trace. The dataset
-    was retrieved from
-    https://www.kaggle.com/c/google-smartphone-decimeter-challenge/data
-
-    References
-    ----------
-    .. [1] Fu, Guoyu Michael, Mohammed Khider, and Frank van Diggelen.
-        "Android Raw GNSS Measurement Datasets for Precise Positioning."
-        Proceedings of the 33rd International Technical Meeting of the
-        Satellite Division of The Institute of Navigation (ION GNSS+
-        2020). 2020.
-    """
-    derived_path = os.path.join(root_path, 'Pixel4_derived.csv')
-    return derived_path
-
-
-@pytest.fixture(name="derived")
-def fixture_load_derived(derived_path):
-    """Load instance of AndroidDerived2021
-
-    Parameters
-    ----------
-    derived_path : pytest.fixture
-        String with location of Android derived measurement file
-
-    Returns
-    -------
-    derived : AndroidDerived2021
-        Instance of AndroidDerived2021 for testing
-    """
-    derived = AndroidDerived2021(derived_path)
-    return derived
-
-def test_residuals_inplace(derived):
+def test_residuals_inplace(derived_2021):
     """Test that solving for residuals doesn't fail
 
     Parameters
     ----------
-    derived : AndroidDerived2021
+    derived_2021 : AndroidDerived2021
         Instance of AndroidDerived2021 for testing.
 
     """
 
-    derived_original = derived.copy()
+    derived_original = derived_2021.copy()
 
-    state_estimate = solve_wls(derived)
+    state_estimate = solve_wls(derived_2021)
 
-    solve_residuals(derived, state_estimate)
+    solve_residuals(derived_2021, state_estimate)
 
     # result should still be a NavData Class instance
-    assert isinstance(derived,type(NavData()))
+    assert isinstance(derived_2021,type(NavData()))
 
     # derived should have one more row but same number of cols
-    assert len(derived.rows) == len(derived_original.rows) + 1
-    assert len(derived) == len(derived_original)
+    assert len(derived_2021.rows) == len(derived_original.rows) + 1
+    assert len(derived_2021) == len(derived_original)
 
     # derived should include new residuals rows but not its copy
-    assert "residuals_m" in derived.rows
+    assert "residuals_m" in derived_2021.rows
     assert "residuals_m" not in derived_original.rows
 
-    assert not np.any(np.isinf(derived["residuals_m"]))
+    assert not np.any(np.isinf(derived_2021["residuals_m"]))
 
     # max is 47.814594604074955
-    assert max(derived["residuals_m"]) < 50.
+    assert max(derived_2021["residuals_m"]) < 50.
 
-def test_residuals(derived):
+def test_residuals(derived_2021):
     """Test that solving for residuals doesn't fail
 
     Parameters
     ----------
-    derived : AndroidDerived2021
+    derived_2021 : AndroidDerived2021
         Instance of AndroidDerived2021 for testing.
 
     """
 
-    state_estimate = solve_wls(derived)
+    state_estimate = solve_wls(derived_2021)
 
-    residuals = solve_residuals(derived, state_estimate, inplace=False)
+    residuals = solve_residuals(derived_2021, state_estimate, inplace=False)
 
     # result should still be a NavData Class instance
     assert isinstance(residuals,type(NavData()))
@@ -129,32 +67,32 @@ def test_residuals(derived):
     # derived should have one more row but same number of cols
     for row in ["residuals_m","gps_millis","gnss_id","sv_id","signal_type"]:
         assert row in residuals.rows
-    assert len(residuals) == len(derived)
+    assert len(residuals) == len(derived_2021)
 
     # derived should not include new residuals row
-    assert "residuals_m" not in derived.rows
+    assert "residuals_m" not in derived_2021.rows
 
     assert not np.any(np.isinf(residuals["residuals_m"]))
 
     # max is 47.814594604074955
     assert max(residuals["residuals_m"]) < 50.
 
-def test_residuals_fails(derived):
+def test_residuals_fails(derived_2021):
     """Test that solving for residuals fails when it should
 
     Parameters
     ----------
-    derived : AndroidDerived2021
+    derived_2021 : AndroidDerived2021
         Instance of AndroidDerived2021 for testing.
 
     """
 
-    state_estimate = solve_wls(derived)
+    state_estimate = solve_wls(derived_2021)
 
     for inplace in [True,False]:
 
         for row in ["gps_millis","corr_pr_m"]:
-            derived_removed = derived.remove(rows=row)
+            derived_removed = derived_2021.remove(rows=row)
             with pytest.raises(KeyError) as excinfo:
                 _ = solve_residuals(derived_removed, state_estimate,
                                     inplace=inplace)
@@ -163,7 +101,7 @@ def test_residuals_fails(derived):
         for row in ["gps_millis"]:
             state_estimate_removed = state_estimate.remove(rows=row)
             with pytest.raises(KeyError) as excinfo:
-                _ = solve_residuals(derived,
+                _ = solve_residuals(derived_2021,
                                     state_estimate_removed,
                                     inplace=inplace)
             assert row in str(excinfo.value)
@@ -176,7 +114,7 @@ def test_residuals_fails(derived):
             error_name = row[:4] + '*' + row[-2:]
             duplicated[new_name] = duplicated[row]
             with pytest.raises(KeyError) as excinfo:
-                _ = solve_residuals(derived,
+                _ = solve_residuals(derived_2021,
                                     duplicated,
                                     inplace=inplace)
             assert error_name in str(excinfo.value)
@@ -185,7 +123,7 @@ def test_residuals_fails(derived):
             state_estimate_removed = state_estimate.remove(rows=row)
             error_name = row[:4] + '*' + row[-2:]
             with pytest.raises(KeyError) as excinfo:
-                _ = solve_residuals(derived,
+                _ = solve_residuals(derived_2021,
                                     state_estimate_removed,
                                     inplace=inplace)
             assert error_name in str(excinfo.value)
